requirements.txt
pycryptodome
requests
secp256k1

